Move UUID back
